+++ geode-core/src/main/java/org/apache/geode/internal/cache/DummyCachePerfStats.java
@@ -37,7 +37,7 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getLoadsCompleted() {
+  public long getLoadsCompleted() {
     return 0;
   }
 
@@ -52,7 +52,7 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getNetloadsCompleted() {
+  public long getNetloadsCompleted() {
     return 0;
   }
 
@@ -67,7 +67,7 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getNetsearchesCompleted() {
+  public long getNetsearchesCompleted() {
     return 0;
   }
 
@@ -92,7 +92,7 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getGetInitialImageKeysReceived() {
+  public long getGetInitialImageKeysReceived() {
     return 0;
   }
 
@@ -107,42 +107,42 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getDestroys() {
+  public long getDestroys() {
     return 0;
   }
 
   @Override
-  public int getCreates() {
+  public long getCreates() {
     return 0;
   }
 
   @Override
-  public int getPuts() {
+  public long getPuts() {
     return 0;
   }
 
   @Override
-  public int getPutAlls() {
+  public long getPutAlls() {
     return 0;
   }
 
   @Override
-  public int getUpdates() {
+  public long getUpdates() {
     return 0;
   }
 
   @Override
-  public int getInvalidates() {
+  public long getInvalidates() {
     return 0;
   }
 
   @Override
-  public int getGets() {
+  public long getGets() {
     return 0;
   }
 
   @Override
-  public int getMisses() {
+  public long getMisses() {
     return 0;
   }
 
@@ -317,12 +317,12 @@ public class DummyCachePerfStats extends CachePerfStats {
   public void endQueryExecution(long executionTime) {}
 
   @Override
-  public int getTxCommits() {
+  public long getTxCommits() {
     return 0;
   }
 
   @Override
-  public int getTxCommitChanges() {
+  public long getTxCommitChanges() {
     return 0;
   }
 
@@ -337,12 +337,12 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getTxFailures() {
+  public long getTxFailures() {
     return 0;
   }
 
   @Override
-  public int getTxFailureChanges() {
+  public long getTxFailureChanges() {
     return 0;
   }
 
@@ -357,12 +357,12 @@ public class DummyCachePerfStats extends CachePerfStats {
   }
 
   @Override
-  public int getTxRollbacks() {
+  public long getTxRollbacks() {
     return 0;
   }
 
   @Override
-  public int getTxRollbackChanges() {
+  public long getTxRollbackChanges() {
     return 0;
   }
